Move SettingsEnhancedMeasurementSwitch to components/settings folder.
